Move initialization later
